+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## dev
 
+* Bug
+  * [Dockerfile] Fixing `azk build` does not support `COPY` in Dockerfile, issue #341.
+
 ## v0.11.0 - (2015-25-03)
 
 * Bug